feat: select refactored
