+++ packages/anu-vue/src/components/select/index.ts
@@ -1,2 +1,2 @@
-export { ASelect } from './ASelect'
+export { default as ASelect } from './ASelect.vue'
